+++ python/semantic_kernel/kernel_extensions/extends_kernel.py
@@ -1,0 +1,10 @@
+# Copyright (c) Microsoft. All rights reserved.
+
+from typing import Protocol
+
+from semantic_kernel.kernel_base import KernelBase
+
+
+class ExtendsKernel(Protocol):
+    def kernel(self) -> KernelBase:
+        ...
